General system stability improvements to enhance the user's experience.
* Changed the data type of angles from double to float, as doubles aren't really required

+++ inc/Variable.hpp
@@ -21,7 +21,7 @@ namespace variable
     struct Target
     {
         uint16_t name;
-        std::vector<double> angles;
+        std::vector<float> angles;
 
     };
 }
